# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AppCloud API",
    "description": "Knowledge graph platform for infrastructure dependency mapping and blast-radius analysis.",
    "version": "1.1.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/integrations/": {
      "get": {
        "summary": "List configured connector instances",
        "tags": [
          "Integrations"
        ],
        "description": "Returns every row from the `integrations` table — ServiceNow, Terraform Cloud, OTel ingest, IaC state backends, etc. Secret fields are decrypted; client UIs are responsible for further obscuring.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a connector instance",
        "tags": [
          "Integrations"
        ],
        "description": "Body shape: `{ type, name, config, enabled?, pollIntervalSeconds? }`. The `type` must match a registered connector (`/connectors`); the connector's `beforeUpsert` runs before validation to fill defaults / auto-generate secrets, and `afterUpsert` runs after persist to sync derived rows. Secret fields are AES-256-GCM-encrypted before INSERT.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "name"
                ],
                "additionalProperties": true,
                "properties": {
                  "type": {
                    "type": "string",
                    "description": "Connector id (e.g. servicenow, terraform-cloud, otel-ingest)"
                  },
                  "name": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "enabled": {
                    "type": "boolean",
                    "default": true
                  },
                  "pollIntervalSeconds": {
                    "type": [
                      "integer",
                      "null"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/integrations/{id}": {
      "get": {
        "summary": "Get one connector instance",
        "tags": [
          "Integrations"
        ],
        "description": "Returns the integration row with config decrypted. Use `/connectors/:type` for the connector's schema and capabilities.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a connector instance (partial)",
        "tags": [
          "Integrations"
        ],
        "description": "Patches `config` (merged + re-encrypted), `enabled`, or `pollIntervalSeconds`. The connector's before/afterUpsert hooks run on each update.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "properties": {
                  "config": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "pollIntervalSeconds": {
                    "type": [
                      "integer",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a connector instance",
        "tags": [
          "Integrations"
        ],
        "description": "Removes the row. Resources the connector previously ingested into Neo4j stay; the next discovery sweep / cleanupStaleNodes will mark them stale.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/integrations/{id}/test": {
      "post": {
        "summary": "Test a connector's healthCheck",
        "tags": [
          "Integrations"
        ],
        "description": "Invokes `spec.healthCheck()` against the live integration. Returns `{ ok, detail }`. Cheap — does not ingest anything.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/integrations/{id}/scan": {
      "post": {
        "summary": "Trigger a one-off ingest pass",
        "tags": [
          "Integrations"
        ],
        "description": "Runs `fetch → normalize → ingest` for the connector and records a `sync_jobs` row with the outcome. Returns the result envelope (resourcesFound / created / updated / warnings).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/integrations/{id}/history": {
      "get": {
        "summary": "Connector instance sync history",
        "tags": [
          "Integrations"
        ],
        "description": "Returns up to 50 of the most recent `sync_jobs` rows for this integration with status, durations, and per-status counts. Use this to debug why an ingest didn't produce the expected nodes.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/connectors/{id}": {
      "get": {
        "summary": "Get one connector's schema",
        "tags": [
          "Connectors"
        ],
        "description": "Returns the full spec for one connector — `id`, `displayName`, `description`, JSON schemas for auth + config, UI metadata, capabilities. Used by the connector-creation UI to drive a form.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://appcloud.local"
    }
  ],
  "security": [
    {
      "ApiKey": []
    }
  ]
}